requirements.txt
coreapi==2.3.3
djangorestframework==3.7.7
Django==1.11.11
factory_boy